create io module
