+++ stateinterpreter/tests/test_md.py
@@ -8,25 +8,26 @@ import pytest
 import stateinterpreter
 import pandas as pd
 
-from stateinterpreter.MD import Loader, plumed_to_pandas
+from stateinterpreter.MD import Loader
+from stateinterpreter.io import load_dataframe
 
 def test_load_colvar():
     """Test loader initialization w/only collective variables"""
 
-    folder = 'data/test-chignolin/'
+    folder = 'stateinterpreter/data/test-chignolin/'
     colvar_file = folder + 'COLVAR'
 
     #load COLVAR from file
     data = Loader(colvar_file, kbt=2.8, stride=20, _DEV=True)
 
     #set COLVAR from dataframe
-    colvar_df = plumed_to_pandas(colvar_file)
+    colvar_df = load_dataframe(colvar_file)
     data2 = Loader(colvar_df, kbt=2.8, stride=20, _DEV=True)
 
 def test_compute_descriptors():
     """Load compute descriptors"""
     
-    folder = 'data/test-chignolin/'
+    folder = 'stateinterpreter/data/test-chignolin/'
     colvar_file = folder + 'COLVAR'
     traj_dict = {
         'trajectory' : folder+'traj.dcd',
@@ -45,17 +46,19 @@ def test_compute_descriptors():
 
     assert data.descriptors.shape == data2.descriptors.shape
 
-#@pytest.mark.skip
+@pytest.mark.skip
 @pytest.mark.parametrize("n_cvs", [1,2])
 def test_identify_states(n_cvs):
     """Identify metastable states based on FES, clustering with differen no. of CVs"""
 
-    folder = 'data/test-chignolin/'
+    folder = 'stateinterpreter/data/test-chignolin/'
     colvar_file = folder + 'COLVAR'
     descr_file = folder + "DESCRIPTORS.csv"
 
     cvs_list = ['deep.node-4','deep.node-3','deep.node-2']
     selected_cvs = cvs_list[:n_cvs]
+    print('cvs: '+" - ".join(selected_cvs))
+
     bounds = [(-1.,1.)]*n_cvs
 
     data = Loader(colvar_file, descr_file, kbt=2.8, stride=1, _DEV=True)
